+++ lizard_auth_client/management/commands/sso_sync_user_organisation_roles.py
@@ -2,25 +2,18 @@ from __future__ import print_function
 from django.contrib.auth import get_user_model
 from django.core.management import call_command
 from django.core.management.base import BaseCommand
-from django.core.management.base import CommandError
-from django.db.utils import IntegrityError
-from lizard_auth_client.client import sso_get_roles_django
-from lizard_auth_client.client import sso_get_user_organisation_roles_django
 from lizard_auth_client.client import sso_sync_user_organisation_roles
-from lizard_auth_client.models import Organisation
-from lizard_auth_client.models import Role
 from lizard_auth_client.models import UserOrganisationRole
-
-import sys
 
 
 txt = {
     'provide_username':
         '[E] Please provide the username for the user you are trying to sync.',
     'winrar':
-        "[+] Succesfully wrote UserOrganisationRole instance(s) to " \
+        "[+] Succesfully wrote UserOrganisationRole instance(s) to "
         "database. Currently we have %i instance(s) for '%s' in the db."
 }
+
 
 class Command(BaseCommand):
     """
@@ -31,7 +24,6 @@ class Command(BaseCommand):
 
     def add_arguments(self, parser):
         parser.add_argument('sso_user', type=str)
-
 
     def handle(self, *args, **options):
 
@@ -45,6 +37,9 @@ class Command(BaseCommand):
             call_command('sso_sync_organisations')
             call_command('sso_sync_user', wanted_username)
             sso_sync_user_organisation_roles(wanted_user)
-            print(txt['winrar'] %
-                (UserOrganisationRole.objects.filter(user=wanted_user).count(),
-                 wanted_user.username))
+            print(
+                txt['winrar'] % (
+                    UserOrganisationRole.objects.filter(
+                        user=wanted_user).count(),
+                    wanted_user.username)
+            )