+++ lizard_auth_client/client.py
@@ -228,6 +228,7 @@ def sso_authenticate_v2(sso_server_api_start_url, sso_key, sso_secret,
 
     """
     try:
+        r = None
         payload = {
             # Identifier for this site
             'iss': sso_key,
@@ -247,9 +248,14 @@ def sso_authenticate_v2(sso_server_api_start_url, sso_key, sso_secret,
             r.raise_for_status()
         return r.json()['user']
 
-    except:
+    except Exception:
+        if r is not None and r.status_code == 403:
+            # Log at INFO so a wrong password doesn't show up in Sentry.
+            logger.info("SSO server returned HTTP_403_FORBIDDEN.")
+            raise AuthenticationFailed
         logger.exception(
-            "Exception occurred while asking SSO to check credentials")
+            "Exception occurred while asking SSO to check credentials"
+        )
         raise